requirements.txt
streamlit
numpy
pillow
tensorflow==2.15.0